moved HighlightOverlay to VisualizationBase
