+++ VisualizationBase/src/overlays/HighlightOverlay.cpp
@@ -24,49 +24,53 @@
 **
 ***********************************************************************************************************************/
 
-#pragma once
+#include "HighlightOverlay.h"
 
-#include "../informationscripting_api.h"
+#include "VisualizationBase/src/shapes/Shape.h"
 
-#include "HighlightOverlayStyle.h"
+#include "VisualizationBase/src/items/TextStyle.h"
+#include "VisualizationBase/src/declarative/DeclarativeItem.hpp"
+#include "VisualizationBase/src/items/EmptyItem.h"
 
-#include "VisualizationBase/src/items/Text.h"
-#include "VisualizationBase/src/overlays/SelectionOverlay.h"
-#include "VisualizationBase/src/declarative/DeclarativeItem.h"
+using namespace Visualization;
 
 namespace Visualization {
-	class EmptyItem;
+
+DEFINE_ITEM_COMMON(HighlightOverlay, "item")
+
+HighlightOverlay::HighlightOverlay(Item* selectedItem, const StyleType* style)
+	: Super{{selectedItem}, style}
+{
+	info_ = new Text{this};
+	info_->setTextFormat(Qt::RichText);
 }
 
-namespace InformationScripting {
-
-class INFORMATIONSCRIPTING_API HighlightOverlay
-		: public Super<Visualization::Overlay<Visualization::DeclarativeItem<HighlightOverlay>>>
+void HighlightOverlay::initializeForms()
 {
-	ITEM_COMMON(HighlightOverlay)
+	auto backgroundElement = item<EmptyItem>(&I::background_, &StyleType::background);
+	auto textItem = item<Text>(&I::info_, &StyleType::info);
+	addForm(backgroundElement);
 
-	public:
-		HighlightOverlay(Item* selectedItem, const StyleType* style = itemStyles().get());
+	addForm((new AnchorLayoutFormElement{})
+			  ->put(TheVCenterOf, textItem, AtCenterOf, backgroundElement)
+			  ->put(TheHCenterOf, textItem, AtCenterOf, backgroundElement));
+}
 
-		static void initializeForms();
-		int determineForm() override;
+int HighlightOverlay::determineForm()
+{
+	return info_->text().size() > 0 ? 1 : 0;
+}
 
-		void setText(const QString& text);
-
-		virtual QColor customShapeColor() const override;
-		void setColor(const QColor& color);
-
-	protected:
-		virtual void updateGeometry(int availableWidth, int availableHeight) override;
-
-	private:
-		QColor color_;
-		Visualization::Text* info_{};
-		Visualization::EmptyItem* background_{};
-};
-
-inline void HighlightOverlay::setText(const QString& text) { info_->setText(text); }
-inline QColor HighlightOverlay::customShapeColor() const { return color_; }
-inline void HighlightOverlay::setColor(const QColor& color) { color_ = color; }
+void HighlightOverlay::updateGeometry(int availableWidth, int availableHeight)
+{
+	background_->setAcceptedMouseButtons(Qt::NoButton);
+	background_->setCustomSize(associatedItem()->widthInScene(), associatedItem()->heightInScene());
+	Super::updateGeometry(availableWidth, availableHeight);
+	if (hasShape())
+	{
+		auto pos = QPointF{(qreal) getShape()->contentLeft(), (qreal) getShape()->contentTop()};
+		setPos(associatedItem()->mapToScene(0, 0) - pos);
+	}
+}
 
 }
